# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/v1/permissions/list:
    post:
      summary: 'Permission list'
      operationId: permissionList
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"data\": [\n{\n\"id\": 1,\n\"name\": \"create user\",\n\"guard_name\": \"web\",\n\"created_at\": \"2021-01-01T00:00:00.000000Z\",\n\"updated_at\": \"2021-01-01T00:00:00.000000Z\"\n},\n},\n\"success\": true,\n\"message\": \"Permission list successfully\"\n}"
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  success: false
                  message: 'Permission list empty'
                properties:
                  data:
                    type: array
                    example: []
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Permission list empty'
      tags:
        - 'Permission APIs'
      security: []
...
  /api/v1/permissions/create:
    post:
      summary: 'Permission create'
      operationId: permissionCreate
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Something went wrong'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Something went wrong'
      tags:
        - 'Permission APIs'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'The name of the permission.'
                  example: 'create permission'
              required:
                - name
      security: []
...